Include symlink targets in recovery_noto-fonts_dep
recovery_noto-fonts_dep is used to generate images that get installed in
recovery.img. This is currently done by make packaging system. The
packaging rules are being migrated to soong, which is more strict about
sandboxing build rules. Some of the src files in
`recovery_noto-fonts_dep` are symlinks to other files in this directory.

In preparation for soong migration, add the symlink targets of
these fonts to `recovery_noto-fonts_dep` filegroup

This should be a noop for now.

Bug: 420665555
Test: m nothing

Change-Id: I3e610a2d764cdc073a7414fc7b965ecbe2091408